backend/requirements/base.txt
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-camel-case==1.4.2
Django==5.2
factory-boy==3.3.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
Pillow==11.2.1
social-auth-app-django==5.4.3
